fixed 独立的点餐桌台管理功能迭代

+++ fuintBackend/fuint-application/src/main/java/com/fuint/common/dto/TableDto.java
@@ -28,7 +28,7 @@ public class TableDto {
     private MtUser memberInfo;
 
     @ApiModelProperty("件数")
-    private Double num;
+    private Integer num;
 
     @ApiModelProperty("金额")
     private BigDecimal amount;
